GenAI_Application/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
curl_cffi==0.6.2
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
docarray==0.40.0
duckduckgo_search==5.2.2
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain==0.1.14
langchain-cli==0.0.21
langchain-community==0.0.31
langchain-core==0.1.39
langchain-experimental==0.0.56
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langserve==0.0.51
langsmith==0.1.40
markdown-it-py==3.0.0
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.16.0
orjson==3.10.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
tenacity==8.2.3
tiktoken==0.6.0
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.9.4
types-requests==2.31.0.20240403
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.23.2
wcwidth==0.2.13
yarl==1.9.4
zipp==3.18.1
